Added dependency injection to login function
The function previously was serving two purposes. It would first search
for the data within a json file. Then it would proceed with login funcionality.
Now the login function has a parameter which requires the data to be passed to the
function first in order for it to function. The dependency in this case is the
argument being injected to the function parameter